fast-api/requirements.txt
fastapi
uvicorn
sqlalchemy
cloud-sql-python-connector[pg8000]
google-cloud-aiplatform